PREDICTOR/pyproject.toml
[project]
name = "glucosense"
version = "0.1.0"
description = "minor project for uni"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.3",
    "google-genai>=1.72.0",
    "gunicorn>=25.3.0",
    "joblib>=1.5.3",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pdfplumber>=0.11.9",
    "pypdf2>=3.0.1",
    "python-multipart>=0.0.26",
    "reportlab>=4.4.10",
    "scikit-learn>=1.8.0",
    "uvicorn>=0.44.0",
]
